+++ properties/Connection/SelectLimit.php
@@ -57,7 +57,7 @@ final class SelectLimit implements Property
     public function ensureHeldBy(Assert $assert, object $connection): object
     {
         $connection(Insert::into(
-            new Name('test'),
+            Name::of('test'),
             Row::of([
                 'id' => $this->uuid,
                 'username' => $this->username,