+++ properties/Connection/UpdateSpecificRow.php
@@ -54,10 +54,10 @@ final class UpdateSpecificRow implements Property
     public function ensureHeldBy(Assert $assert, object $connection): object
     {
         $insert = Query\MultipleInsert::into(
-            new Table\Name('test'),
-            new Column\Name('id'),
-            new Column\Name('username'),
-            new Column\Name('registerNumber'),
+            Table\Name::of('test'),
+            Column\Name::of('id'),
+            Column\Name::of('username'),
+            Column\Name::of('registerNumber'),
         );
         $connection($insert(Sequence::of(
             Row::of([
@@ -73,7 +73,7 @@ final class UpdateSpecificRow implements Property
         )));
 
         $update = Query\Update::set(
-            new Table\Name('test'),
+            Table\Name::of('test'),
             Row::of(['registerNumber' => 24]),
         );
         $update = $update->where(new class($this->uuid1) implements Comparator {